requirements.txt
asyncio
websockets
opencv-python
python-dotenv
elevenlabs
soundfile # Read sound file
sounddevice #change sound speed
gtts #text to sound